application/external_apps/mcp/requirements.txt
mcp
requests
python-dotenv
